% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_comparison.R
\name{reg_comparison}
\alias{reg_comparison}
\title{Compare Regression Models}
\usage{
reg_comparison(reg1, reg2)
}
\arguments{
\item{reg1}{The first regression model.}

\item{reg2}{The second regression model.}
}
\value{
A data frame containing the results of the comparison.
}
\description{
This function compares the coefficients of two regression models and performs a significance test.
}
\examples{
library(knitr)
reg1 <- lm(y ~ x, data = data1)
reg2 <- lm(y ~ x, data = data2)
comparison <- reg_comparison(reg1, reg2)
print(comparison)
}
